requirements.txt
django-annoying==0.10.3
django-markdown-app==0.9.2
django-taggit==0.22.1
django-hitcount==1.2.2
django-markdownx
django-mathfilters